RoleStore Implemented, Role Repository, Service Models and DTOs added

+++ WebApplicationClient/IdentityInfrastructure/CustomIdentity.cs
@@ -12,15 +12,18 @@ using Microsoft.Owin;
 
 namespace WebApplicationClient.IdentityInfrastructure
 {
-    public class CustomUserManager: UserManager<User, int>
+    public class CustomUserManager : UserManager<User, int>
     {
-        public CustomUserManager(IUserStore<User, int> store):base(store)
+        public CustomUserManager(IUserStore<User, int> store) : base(store)
         {
 
         }
         public static CustomUserManager Create(IdentityFactoryOptions<CustomUserManager> options, IOwinContext context)
         {
-            return new CustomUserManager(new CustomUserStore((IUserInfoService)System.Web.Mvc.DependencyResolver.Current.GetService(typeof(IUserInfoService))));
+            return new CustomUserManager(
+                new CustomUserStore(
+                    (IUserInfoService)System.Web.Mvc.DependencyResolver.Current.GetService(typeof(IUserInfoService)),
+                    (IRoleService)System.Web.Mvc.DependencyResolver.Current.GetService(typeof(IRoleService))));
         }
 
     }
@@ -35,7 +38,8 @@ namespace WebApplicationClient.IdentityInfrastructure
                 Id = userDTO.Id,
                 UserName = userDTO.UserName,
                 PasswordHash = userDTO.PasswordHash,
-                Email = userDTO.Email
+                Email = userDTO.Email,
+                Roles = userDTO.Roles
             };
         }
 
@@ -48,23 +52,33 @@ namespace WebApplicationClient.IdentityInfrastructure
         }
     }
 
-    public class CustomUserStore : IUserStore<User, int>, IUserPasswordStore<User, int>
+    public class CustomUserStore : IUserStore<User, int>, IUserPasswordStore<User, int>, IUserRoleStore<User, int>
     {
         IUserInfoService userInfoService;
+        IRoleService roleService;
 
-        public CustomUserStore(IUserInfoService userInfoService)
+        public CustomUserStore(IUserInfoService userInfoService, IRoleService roleService)
         {
             this.userInfoService = userInfoService;
+            this.roleService = roleService;
         }
 
         Task IUserStore<User, int>.CreateAsync(User user)
         {
+            if (user == null)
+            {
+                throw new ArgumentNullException("user");
+            }
             this.userInfoService.Insert(user);
             return Task.FromResult(0);
         }
 
         Task IUserStore<User, int>.DeleteAsync(User user)
         {
+            if (user == null)
+            {
+                throw new ArgumentNullException("user");
+            }
             this.userInfoService.Delete(user.Id);
             return Task.FromResult(0);
         }
@@ -76,18 +90,26 @@ namespace WebApplicationClient.IdentityInfrastructure
 
         Task<User> IUserStore<User, int>.FindByNameAsync(string userName)
         {
+            if (string.IsNullOrEmpty(userName))
+            {
+                throw new ArgumentNullException("userName");
+            }
             return Task.FromResult(User.FromDTO(this.userInfoService.GetByUsername(userName)));
         }
 
         Task IUserStore<User, int>.UpdateAsync(User user)
         {
+            if (user == null)
+            {
+                throw new ArgumentNullException("user");
+            }
             this.userInfoService.Update(user);
             return Task.FromResult(0);
         }
 
         Task IUserPasswordStore<User, int>.SetPasswordHashAsync(User user, string passwordHash)
         {
-            if(user == null)
+            if (user == null)
             {
                 throw new ArgumentException("user");
             }
@@ -97,6 +119,10 @@ namespace WebApplicationClient.IdentityInfrastructure
 
         Task<string> IUserPasswordStore<User, int>.GetPasswordHashAsync(User user)
         {
+            if (user == null)
+            {
+                throw new ArgumentNullException("user");
+            }
             return Task.FromResult(user.PasswordHash);
         }
 
@@ -144,6 +170,81 @@ namespace WebApplicationClient.IdentityInfrastructure
             Dispose(true);
             // TODO: uncomment the following line if the finalizer is overridden above.
             // GC.SuppressFinalize(this);
+        }
+
+        Task IUserRoleStore<User, int>.AddToRoleAsync(User user, string roleName)
+        {
+            if (user == null)
+            {
+                throw new ArgumentNullException("user");
+            }
+
+            if (string.IsNullOrEmpty(roleName))
+            {
+                throw new ArgumentException("roleName");
+            }
+
+            var role = this.roleService.GetByName(roleName);
+            if(user.Roles == null)
+            {
+                user.Roles = new List<RoleDTO>();
+            }
+            if (role != null && user.Roles.All(r => r.Id != role.Id))
+            {
+                user.Roles.Add(role);
+            }
+
+            return Task.FromResult(0);
+        }
+
+        Task IUserRoleStore<User, int>.RemoveFromRoleAsync(User user, string roleName)
+        {
+            if (user == null)
+            {
+                throw new ArgumentNullException("user");
+            }
+            if (String.IsNullOrWhiteSpace(roleName))
+            {
+                throw new ArgumentNullException("roleName");
+            }
+            var role = this.roleService.GetByName(roleName);
+            if (role != null && user.Roles.Any(r => r.Id == role.Id))
+            {
+                user.Roles.Remove(role);
+                this.userInfoService.Update(user);
+            }
+            return Task.FromResult(0);
+        }
+
+        Task<IList<string>> IUserRoleStore<User, int>.GetRolesAsync(User user)
+        {
+            if (user == null)
+            {
+                throw new ArgumentNullException("user");
+            }
+
+            if (user.Roles == null) return Task.FromResult<IList<string>>(new List<string>());
+
+            var roles = user.Roles.Select(r => r.Name).ToList();
+
+            return Task.FromResult<IList<string>>(roles);
+        }
+
+        Task<bool> IUserRoleStore<User, int>.IsInRoleAsync(User user, string roleName)
+        {
+            if (user == null)
+            {
+                throw new ArgumentNullException("user");
+            }
+
+            if (string.IsNullOrEmpty(roleName))
+            {
+                throw new ArgumentNullException("roleName");
+            }
+
+            var isInRole = user.Roles.Any(r => r.Name == roleName);
+
+            return Task.FromResult(isInRole);
         }
         #endregion
